added context to delete comment

+++ server/schemas/resolvers.js
@@ -63,8 +63,12 @@ const resolvers = {
       }
       throw new AuthenticationError("That's not your comment!");
     },
-    deleteComment: async (parent, { commentId }) => {
-      return await Comment.findByIdAndDelete({commentId}) 
+    deleteComment: async (parent, { commentId }, context) => {
+      const deletedComment = Comment.find({_id: commentId});
+      if (context.user._id === deleteddComment.userId) {
+        return await Comment.findByIdAndDelete({commentId}) 
+      }
+      throw new AuthenticationError("That's not your comment!");
     },
     addTwit: async (parent, { twitText }, context) => {
       if (context.user) {
